+++ Infra/Migrations/20260107193224_InitialCreate.cs
@@ -1,0 +1,159 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace MyVocaList.Infra.Migrations
+{
+    /// <inheritdoc />
+    public partial class InitialCreate : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "People",
+                columns: table => new
+                {
+                    Id = table.Column<int>(type: "INTEGER", nullable: false)
+                        .Annotation("Sqlite:Autoincrement", true),
+                    FullName = table.Column<string>(type: "TEXT", maxLength: 250, nullable: false, collation: "NOCASE_NOACCENT"),
+                    FullNameNormalized = table.Column<string>(type: "TEXT", maxLength: 250, nullable: true, collation: "NOCASE_NOACCENT"),
+                    Participations = table.Column<int>(type: "INTEGER", nullable: false, defaultValue: 0),
+                    Absences = table.Column<int>(type: "INTEGER", nullable: false, defaultValue: 0),
+                    BirthdayDayMonth = table.Column<string>(type: "TEXT", nullable: false, collation: "NOCASE_NOACCENT"),
+                    Email = table.Column<string>(type: "TEXT", nullable: false, collation: "NOCASE_NOACCENT")
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_People", x => x.Id);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "SystemConfigurations",
+                columns: table => new
+                {
+                    Id = table.Column<int>(type: "INTEGER", nullable: false)
+                        .Annotation("Sqlite:Autoincrement", true),
+                    Key = table.Column<string>(type: "varchar(50)", maxLength: 50, nullable: false, collation: "NOCASE_NOACCENT"),
+                    Value = table.Column<string>(type: "varchar(200)", maxLength: 200, nullable: false, collation: "NOCASE_NOACCENT")
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_SystemConfigurations", x => x.Id);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Venues",
+                columns: table => new
+                {
+                    Id = table.Column<int>(type: "INTEGER", nullable: false)
+                        .Annotation("Sqlite:Autoincrement", true),
+                    Name = table.Column<string>(type: "TEXT", maxLength: 30, nullable: false, collation: "NOCASE_NOACCENT")
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Venues", x => x.Id);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "Events",
+                columns: table => new
+                {
+                    Id = table.Column<int>(type: "INTEGER", nullable: false)
+                        .Annotation("Sqlite:Autoincrement", true),
+                    VenueId = table.Column<int>(type: "INTEGER", nullable: false),
+                    EventDate = table.Column<DateTime>(type: "TEXT", nullable: false),
+                    EventName = table.Column<string>(type: "nvarchar(50)", maxLength: 50, nullable: false, collation: "NOCASE_NOACCENT"),
+                    QueueActive = table.Column<bool>(type: "INTEGER", nullable: false, defaultValue: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Events", x => x.Id);
+                    table.ForeignKey(
+                        name: "FK_Events_Venues_VenueId",
+                        column: x => x.VenueId,
+                        principalTable: "Venues",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateTable(
+                name: "EventParticipations",
+                columns: table => new
+                {
+                    Id = table.Column<int>(type: "INTEGER", nullable: false)
+                        .Annotation("Sqlite:Autoincrement", true),
+                    PersonId = table.Column<int>(type: "INTEGER", nullable: false),
+                    EventId = table.Column<int>(type: "INTEGER", nullable: false),
+                    Timestamp = table.Column<DateTime>(type: "TEXT", nullable: false),
+                    Status = table.Column<int>(type: "INTEGER", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_EventParticipations", x => x.Id);
+                    table.ForeignKey(
+                        name: "FK_EventParticipations_Events_EventId",
+                        column: x => x.EventId,
+                        principalTable: "Events",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                    table.ForeignKey(
+                        name: "FK_EventParticipations_People_PersonId",
+                        column: x => x.PersonId,
+                        principalTable: "People",
+                        principalColumn: "Id",
+                        onDelete: ReferentialAction.Restrict);
+                });
+
+            migrationBuilder.CreateIndex(
+                name: "IX_EventParticipations_EventId",
+                table: "EventParticipations",
+                column: "EventId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_EventParticipations_PersonId",
+                table: "EventParticipations",
+                column: "PersonId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_Events_VenueId",
+                table: "Events",
+                column: "VenueId");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_People_BirthdayDayMonth",
+                table: "People",
+                column: "BirthdayDayMonth");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_People_Email",
+                table: "People",
+                column: "Email");
+
+            migrationBuilder.CreateIndex(
+                name: "IX_People_FullNameNormalized",
+                table: "People",
+                column: "FullNameNormalized");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "EventParticipations");
+
+            migrationBuilder.DropTable(
+                name: "SystemConfigurations");
+
+            migrationBuilder.DropTable(
+                name: "Events");
+
+            migrationBuilder.DropTable(
+                name: "People");
+
+            migrationBuilder.DropTable(
+                name: "Venues");
+        }
+    }
+}